engine: extract esm/layers/{layer_ui,object3d}.js
